refactor: centralize profile services
